Re-architect HTML/CSS/SCSS/JS
The architecture of the way we were using SASS/SCSS and
Bootstrap did not take advantage of the fact that Bootstrap
is componentized and overwritable.  Most of our definitions
should be created as partials and then imported without
compilation so that it can use the varibles and mixins
defined within each library imported.

There were a lot of CSS classes created to implement
the design styles that were not needed at all.  Most everything
we need is already in Bootstrap, and a few CSS tweaks will get
you the layout you want.

For the new architecture:

application is what everything is compiled into, which
imports the bootstrap-theme, fontawesome, and the
myusa (myusa-specific) stylesheets

bootstrap-theme includes partials that override bootstrap:
- colors -- put all colors in here
- fonts -- all the font definitions
- elements in bootstrap-theme directly override bootstrap before it is loaded
- bootstrap-overrides overwrites css classes after bootstrap is imported
for elements that don't have variables that can be manipulated in SCSS

new.html for authorizations has been updated to show how to use
bootstrap effectively with less code and less css class definitions

the javascript has been updated to use the new class definitions

oauth_helper is updated to provide either an input box or text
depending on whether the value is already set.