Put the Devise return path in token auth link
If user is sent to the MyUSA login form while they are in the middle of something, we need to be able to keep the return path. So, pass it as an option to the token generation method and then stash it in the email link that we send to users. Then, when the token is authenticated, look for the return path, and if present redirect back to it.