Update MyUSA button and nav pils
The MyUSA button was just using btn-primary and not the actual
MyUSA button specified in the wireframes.  Having consistency
in our social buttons is important.

Create new `nav-4` and `nav-3` classes for the headers so that
they don't move around when hovered.  The problem was that
hovering made them bold, which changed their positioning.  Using
static % positioning fixes this optical challenge.

Revert profile link until the profile implementation is merged.