Partner App Logout
Supports a redirect after logout.  By passing "continue=<url>" as a logout
parameter, an app can redirect the user back to any sub-url of their site.  In
order to be allowed, the redirect URL must be a hostname that ends with the
same string as an application's URL, for an application approved by the user.