Add remote_ip and created_at to UserAction model
Add remote_ip and created_at to UserAction model. Also add a 'Sweeper' to the
audit classes so that we can intercept a UserAction.create and add remote_ip
(and current_user) to the record. Also add Warden hooks for after_set_user and
before_logout so that we can audit sign_in and sign_out events.