Add legal documentation to MyUSA & link from footer
This addresses #49 and #108.  It posts the legal docs
as markdown, and then processes them into one legal.md file
using a rake task.  From there, aglio turns the markdown
into html, and puts it in /public/legal.html.

aglio part of our rails views, but we're leaving that for
another day.