Move API docs to Rails ERB
By having Aglio generate an ERB partial, we can render
the API docs from the same context as other pages in
the application.

Move the route for the developer api docs to /developer
since this will be the only page there.

Update the rake task to generate the _developer partial.

Add Aglio stylings to _aglio.css.scss -- these are
needed to deal with the classes that Aglio expects
to exist in the CSS.

A few minor nits:

- Had to revert back some of my aglio template changes
  due to an aglio bug. This keeps us from rounding the corners
  on the nav elements on the left
- If someone has a really low res desktop, the nav items on the
  left may not all be visible.  C'est la vie.