Add a responsive nav bar for secondary nav
When using on tablets or phones, the secondary nav was
hidden by default (this is an aglio thing).  It makes
sense for long documents (like the legal text),
but in the app you still need to be able to navigate
to the different sub-nav sections.

This method creates a **different** nav html for
small devices (xs and sm) than large devices (md and lg).
This is probably NOT the best way to do this, but it works
for now enough to demo and test.  We can cleanup this
HTML+CSS to be better when we decide on a nav design.