Add select placeholders and redirect to target
When you submit from the 'Personal Information' page, you
should return to that page.  When you submit from the
'Additional Information' page, you should return to that page
(instead of the previous behavior of returning to `/profile`)
Add a hidden `target` field and redirect to that if present

Add placeholders for the select boxes, which may be different
in profile than in the authorization screen due to the
layout and formatting (and context).

Move some more strings into the locale file.