Responsive spacing issues; section centering
Reduced the arrows in the steps and put them near the numbers.
This is instead of the big arrows. Centered the section with
the numbers using bootstrap columns.

Made the width of the login block constrained across different
screen resolutions, and used the bootstrap center-block option
to center it on the page.

Changes requested by @dutchashell