Provide a different message for logged in users
When viewing the homepage after logging in, it asks you to
log in again.  Now that doesn't make much sense.  Instead,
let people go to their profile or applications.

This is a temporary change until the rest of the dashboard
app is built out, but provides a better experience than
we had before and is "good enough" for testing.

@dutchashell reviewed and is ok with this.