Update content to comply with Privacy Act requirements
The Privacy Act requires us to talk about our information
collection DIRECTLY on the page that collects that information.
While it has the best intentions, it really just results in a
lot of small type.  A future action for us is to user test
what do users want to know, and remove these silly disclaimers.

In order to be compliant, the following items have changed:

- Rename to "Terms of Service" everywhere
- Add Privacy Act notice to profile edit, additional info edit,
  and authorization page
- Add an explicit line before "Allow Access" that tells the
  user their information will be shared (Privacy Act requirement)
- Use the Privacy Act notice provided by GSA OGC