Replace MySQL time methods with Rails methods
Rails provides methods that allow you to compare a Date field to a
particular time so that you don't have to use methods that are specific
to your DB.

The advantage is that you are not locked to a specific database, and you
don't have to force users of this app to change their global MySQL
settings.

While I was in there, I also corrected some Rubocop offenses and did
some small refactoring of the Feedback model to improve performance and
reduce duplication. Namely, I added a scope to look up the `remote_ip`,
and I used `size` instead of `count` because the former is faster.